fix(lint): remove redundant i32 cast on RawFd in TestCallback
clippy::unnecessary_cast: RawFd is already i32 on Linux/Android.

+++ native/rust/crates/ripdpi-native-protect/src/lib.rs
@@ -50,7 +50,7 @@ mod tests {
 
     impl ProtectCallback for TestCallback {
         fn protect(&self, fd: RawFd) -> io::Result<()> {
-            self.last_fd.store(fd as i32, Ordering::Relaxed);
+            self.last_fd.store(fd, Ordering::Relaxed);
             Ok(())
         }
     }
